+++ datarep6/addf04.cc
@@ -1,6 +1,6 @@
 extern "C" {
 
-char* add(char* a, int b) {
+char* add(char* a, long b) {
     return &a[b];
 }
